functions/streaming_api_function/requirements.txt
functions-framework==3.0.0
google-cloud-firestore==0.30.1
google-cloud-bigquery==1.8.1
google-cloud-pubsub==0.34.0
pytz==2018.7
gcsfs==2022.2.0
fsspec==2022.2.0
flask==2.0.2